speedtrap-web-services/py-mongo-api/requirements.txt
flask
gunicorn
pymongo
python-dotenv
requests
authlib
six
